Add files via upload
PGP Data Science Engineering
Introduction to Programming – Mini Project:
	Problem 1:
Write a program to calculate area of shapes. Your program should be capable of calculating the area of a square, rectangle, triangle and a circle. The user should be presented with options to select the shape. Based on which shape is chosed by the user, the program should ask for the appropriate input and print the resulting area on the screen.
----------------------------------------------------------------------------------------------------------------
When the program is run, the screen should display something like this:

	Which shape would you like to calculate the area for? Please enter the option number-
	Square
	Rectangle
	Triangle
	Circle
	Enter Option: _

	Say the user enters the option 1.

	Please enter the length of a side: _

	If the user enters a value of 5. The output should be:

	The area of the square is 25

This program should indicate that the input is invalid if the user enters a character instead of a number as input. For instance, if the user enters a value of ‘a’ instead of 5 in the previous example the program should prompt:

Invalid input, please enter a number: _